==> docker-eda/03-load-data-to-mongo-final.ipynb <==
# Install Packages

%matplotlib inline

from __future__ import print_function
from datetime import datetime
import io
import json
import os
import pickle

# Reload the Metadata

If running from a Docker container, make sure to run `aws configure` manually to configure your credentials, or it won't be able to upload to or download from the bucket.

s3_bucket = 'mdang.w210'

# If we don't have any of the cached data files, download them from S3

if not os.path.isdir('iati'):
    !aws s3 cp s3://{s3_bucket}/iati.tar.gz .
    !tar -zxf iati.tar.gz

if not os.path.isfile('activities_metadata.json'):
    !aws s3 cp s3://{s3_bucket}/activities_metadata.json.gz .
    !gunzip activities_metadata.json.gz

if not os.path.isdir('iati-orgs'):
    !aws s3 cp s3://{s3_bucket}/iati-orgs.tar.gz .
    !tar -zxf iati-orgs.tar.gz

if not os.path.isfile('organization_metadata.json'):
    !aws s3 cp s3://{s3_bucket}/organization_metadata.json.gz .
    !gunzip organization_metadata.json.gz

## Activities

from mongodb_iati import activities, activities_metadata, transactions, activity_failures, activity_xml_failures, import_activity_document

activity_failures.clear()
activity_xml_failures.clear()

file_names = [
    '%s/%s' % ('iati', file_name)
        for file_name in os.listdir('iati')
            if file_name.endswith('.xml')
]

for num_finished, file_name in enumerate(file_names):
    if num_finished % 100 == 0:
        failure_count = len(activity_failures) + len(activity_xml_failures)
        print('%s Processed %d of %d (%d errors)' % (datetime.now(), num_finished, len(file_names), failure_count))

    with open(file_name) as f:
        import_activity_document(f)

print('%s Processed %d of %d' % (datetime.now(), num_finished, len(file_names)))

print(activities.count())
print(transactions.count())
print(activities_metadata.count())

activities.find_one()

activities_metadata.find_one()

transactions.find_one()

## Organizations

from mongodb_iati import organizations, organizations_metadata, import_organization_document, organization_failures, organization_xml_failures

file_names = [
    '%s/%s' % ('iati-orgs', file_name)
        for file_name in os.listdir('iati-orgs')
            if file_name.endswith('.xml')
]

for num_finished, file_name in enumerate(file_names):
    if num_finished % 100 == 0:
        failure_count = len(organization_failures) + len(organization_xml_failures)
        print('%s Processed %d of %d' % (datetime.now(), num_finished, len(file_names)))

    with open(file_name) as f:
        import_organization_document(f)

print(organizations.count())
print(organizations_metadata.count())
print(len(organization_xml_failures))

organizations.find_one()

organizations_metadata.find_one()

with open('activity_xml_failures', 'wb') as f:
    pickle.dump(activity_xml_failures, f)

with open('activity_failures', 'wb') as f:
    pickle.dump(activity_failures, f)

with open('organization_xml_failures', 'wb') as f:
    pickle.dump(organization_xml_failures, f)

with open('organization_failures', 'wb') as f:
    pickle.dump(organization_xml_failures, f)

# Investigating/Fixing Errors

def check_errors(folder, metadata, failures):
    metadata_titles = {}
    metadata_resources = {}

    for metadata_item in metadata:
        if 'isopen' not in metadata_item:
            continue

        if not metadata_item['isopen']:
            continue

        title = metadata_item['title']
        short_name = metadata_item['name']

        resources = metadata_item['resources']
        resource_urls = [x for x in set([resource_item['url'] for resource_item in resources])]

        metadata_titles[short_name] = title
        metadata_resources[short_name] = resource_urls

    for short_name, resource_urls in metadata_resources.items():
        iati_filename = folder + '/' + short_name + '.xml'

        if iati_filename not in failures:
            continue

        print('Processing %s failed: %s' % (short_name, failures[iati_filename]))
        print('\n'.join(resource_urls))

## Activities

with io.open('activities_metadata.json', 'r', encoding = 'utf8') as f:
    activity_metadata_json = json.load(f)

print(len(activity_xml_failures))
check_errors('iati', activity_metadata_json, activity_xml_failures)

print(len(activity_failures))
check_errors('iati', activity_metadata_json, activity_failures)

## Organizations

with io.open('organization_metadata.json', 'r', encoding = 'utf8') as f:
    organization_metadata_json = json.load(f)

print(len(organization_xml_failures))
check_errors('iati-orgs', organization_metadata_json, organization_xml_failures)

print(len(organization_failures))
check_errors('iati-orgs', organization_metadata_json, organization_failures)
