==> docker-eda/04-load-data-to-picklegraph.ipynb <==
# Install Packages

%matplotlib inline

from __future__ import print_function
from datetime import datetime
import pickle
import pymongo
import six

# Data Check

If running from a Docker container, make sure to run `aws configure` manually to configure your credentials, or it won't be able to upload to or download from the bucket.

s3_bucket = 'mdang.w210'

conn=pymongo.MongoClient('mongodb', 27017)

conn.database_names()

db = conn.iati

activities=db.activities
activities_metadata=db.activities_metadata
transactions=db.transactions
organizations=db.organizations
organizations_metadata=db.organizations_metadata

db.collection_names()

print(activities_metadata.count(), activities.count())

print(transactions.count())

print(organizations_metadata.count(), organizations.count())

# Add Activities

from parse_activity import add_activity_node, publisher_nodes, organization_nodes, activity_nodes, activity_relationships

## Load Activities and Organizations

activities_count = activities.count()

for num_finished, activity in enumerate(activities.find()):
    if num_finished % 10000 == 0:
        print('%s Processed %d of %d' % (datetime.now(), num_finished, activities_count))

    add_activity_node(activity)

print('%s Processed %d of %d' % (datetime.now(), activities_count, activities_count))

## Check Results

len(publisher_nodes)

len(organization_nodes)

len(activity_nodes)

len(activity_relationships)

## Confirm the Keys in Our Nodes

node_keys = set()

for key, node in activity_nodes.items():
    try:
        node_keys |= set(node.keys())
    except:
        break

node_keys

## Store as Pickled Files

with open('graph_publisher_nodes.pickle', 'wb') as f:
    pickle.dump(publisher_nodes, f)

with open('graph_organization_nodes.pickle', 'wb') as f:
    pickle.dump(organization_nodes, f)

with open('graph_activity_nodes.pickle', 'wb') as f:
    pickle.dump(activity_nodes, f)

with open('graph_activity_relationships.pickle', 'wb') as f:
    pickle.dump(activity_relationships, f)

# Backup Files

Because commands starting with `!` create a forked process, it's possible that the commands below might fail. To work around that possibility, you may want to restart the kernel to free up memory and only run the subsequent commands after that.

!rm -f graph_pickle.tar.gz
!tar -cf graph_pickle.tar graph_*.pickle
!gzip graph_pickle.tar

s3_bucket = 'mdang.w210'

!aws s3 cp graph_pickle.tar.gz s3://{s3_bucket}/ --acl public-read
