==> docker-eda/05-check-organization-names.ipynb <==
# Install Packages

%matplotlib inline

from __future__ import print_function
from collections import defaultdict
from datetime import datetime
import pandas
import pickle
from pprint import PrettyPrinter
import pymongo

printer = PrettyPrinter()

# Data Check

conn=pymongo.MongoClient('mongodb', 27017)

conn.database_names()

db = conn.iati

activities=db.activities
activities_metadata=db.activities_metadata
organizations = db.organizations

db.collection_names()

print(activities_metadata.count(), activities.count())

# Cache in Memory

all_metadata = {}
metadata_count = activities_metadata.count()

print(datetime.now(), 'Started processing')

for num, metadata_item in enumerate(activities_metadata.find()):
    all_metadata[metadata_item['@w210-key']] = metadata_item

print(datetime.now(), 'Finished processing')

len(all_metadata)

# Store URLs from Metadata

iati_urls = {}

def add_from_activity_metadata(filename, metadata):
    if 'organization' not in metadata:
        return

    organization = metadata['organization']

    if organization is None:
        return

    urlname = organization['name']
    title = organization['title']

    iati_urls[filename] = (urlname, title)

iati_urls.clear()

for filename, metadata in all_metadata.items():
    add_from_activity_metadata(filename, metadata)

len(iati_urls)

So we see that about 1 organization doesn't have a clean URL we can easily derive. We'll attempt to add it from the activity.

def add_from_activity(filename, activity):
    if 'participating-org' not in activity:
        return

    organization = activity['participating-org']

    if organization is None:
        return

    if type(organization) == list:
        for organization_item in organization:
            if '@role' in organization_item and organization_item['@role'] == 'Implementing':
                organization = organization_item
                break
   
    if '@ref' not in organization:
        return

    organization_ref = organization['@ref']
    title = organization['#text']

    if title is not None:
        iati_urls[filename] = (organization_ref, title)

Let's take a look at what's missing and see how we fill it in.

for filename in (set(all_metadata.keys()) - set(iati_urls.keys())):
    for activity in activities.find({'@w210-key': filename}):
        add_from_activity(filename, activity)

        if filename in iati_urls:
            print(filename, iati_urls[filename])
        else:
            print(filename, None)

len(iati_urls)

It looks like we were able to drive some additional metadata for that missing organization. Let's go ahead and pickle this file.

with open('iati_urls.pickle', 'wb') as f:
    pickle.dump(iati_urls, f)

## Cache Activities

Bring all the activities out of MongoDB and into memory so that we can process them much faster. Iterating over the MongoDB collection takes minutes. Iterating over an in-memory version containing only the fields we're interested in takes seconds.

However, we need a lot of memory to fit everything, so we'll only include a few of the fields which might be useful when constructing the graph.

* http://iatistandard.org/202/activity-standard/elements/

iati_urls_by_ref = {}

def add_from_activity(item):
    if 'reporting-org' not in activity:
        return

    filename = activity['@w210-key']

    if filename in iati_urls:
        iati_name = iati_urls[filename][0]
    else:
        iati_name = filename[0:filename.find('-')]

    iati_url = 'https://www.iatiregistry.org/publisher/%s' % iati_name

    organization = activity['reporting-org']

    if '@ref' not in organization:
        return

    organization_ref = organization['@ref']

    if organization_ref not in iati_urls_by_ref:
        iati_urls_by_ref[organization_ref] = set([iati_url])
    else:
        iati_urls_by_ref[organization_ref].add(iati_url)

iati_urls_by_ref = {}

print(datetime.now(), 'Started processing')

for activity in activities.find():
    add_from_activity(activity)

print(datetime.now(), 'Finished processing')

Let's pickle the file.

with open('iati_urls_by_ref.pickle', 'wb') as f:
    pickle.dump(iati_urls_by_ref, f)

# Check Fields (Summary)

Let's get a sense of how populated each field is in the data.

fields = set([
    '@w210-key',
    'iati-identifier',
    'other-identifier',
    'reporting-org',
    'participating-org',
    'activity-status',
    'recipient-country',
    'recipient-region',
    'location',
    'sector',
    'transaction'
])

# Count how many times each of the keys of interest appears
# in our activity data set.

field_counts = { key: 0 for key in fields }
field_keys = field_counts.keys()

print(datetime.now(), 'Started processing')

for activity in activities.find():
    for key in field_keys:
        if key in activity:
            field_counts[key] += 1

print(datetime.now(), 'Finished processing')

field_counts

Based on these counts, we'll prefer anything where we can:

* Include the file name
* Include the participating organization
* Include the reporting organization
* Include the transaction information

# Check Fields (One Record)

activity = activities.find_one({ key: {'$ne': None } for key in fields })

activity['iati-identifier']

# http://iatistandard.org/202/activity-standard/iati-activities/iati-activity/sector/
# Vocabulary: http://iatistandard.org/202/codelists/SectorVocabulary/

activity['sector']

# Ref: http://iatistandard.org/202/activity-standard/iati-activities/iati-activity/participating-org/
# Type: http://iatistandard.org/202/codelists/OrganisationType/

activity['reporting-org']

# Ref: http://iatistandard.org/202/activity-standard/iati-activities/iati-activity/participating-org/
# Role: http://iatistandard.org/202/codelists/OrganisationRole/
# Type: http://iatistandard.org/202/codelists/OrganisationType/

activity['participating-org']

# Code: http://iatistandard.org/202/codelists/ActivityStatus/

activity['activity-status']

# Code: http://iatistandard.org/202/codelists/Country/

activity['recipient-country']

# FinanceType: http://iatistandard.org/202/codelists/FinanceType/
# TransactionType: http://iatistandard.org/202/codelists/TransactionType/

activity['transaction']

# Check Data

def flatten(item_list):
    if type(item_list) != list:
        return [item_list]

    return_value = []

    # Recursively flatten lists of lists

    for item in item_list:
        if type(item) == list:
            return_value += flatten(item)
        else:
            return_value.append(item)

    return return_value

def get_text(element, attribute):

    if element is None:
        return None

    if attribute not in element:
        return None

    try:
        value = element[attribute]
    except Exception as e:
        print(element, attribute)
        raise e

    if type(value) == dict:
        if '#text' in value:
            return value['#text']

        return None

    if type(value) != list:
        return value

    return [
        item['#text'] if type(item) == dict and '#text' in item else
            None if type(item) == dict else item
        for item in value
    ]

def get_node_list(parent, field_keys):
    value = parent

    for key in field_keys:

        # If we have a dictionary, we simply access the attribute

        if type(value) == dict:
            if key not in value:
                return []

            value = value[key]
            continue

        # If we have something that is neither a dict nor a list, we
        # cannot navigate further down the JSON object, so we were
        # unable to find what we needed.

        if type(value) != list:
            return []

        # If we have a list, then we'll check the key in each element
        # of the list.

        value = [ item[key] for item in value if item is not None and key in item ]

    if value is None:
        return []

    if type(value) != list:
        value = [value]

    return value

def get_value(activity, field_keys):
    node_list = get_node_list(activity, field_keys)

    return_value = []

    for node in node_list:
        ref = get_text(node, '@ref')
        narrative = get_text(node, 'narrative')

        if narrative is not None:
            return_value.append({
                'ref': ref, 'narrative': narrative
            })

        narrative = get_text(node, '@narrative')

        if narrative is not None:
            return_value.append({
                'ref': ref, 'narrative': narrative
            })

    return return_value

def get_values(field_path):
    field_keys = field_path.split('.')

    return_values = []

    for activity in activities.find():
        new_values = get_value(activity, field_keys)

        for value in new_values:
            value['file_id'] = activity['@w210-key']

        return_values += new_values

    return return_values

def get_new_version_values(element_name):
    return pandas.DataFrame(get_values(element_name))

## Reporting Organizations

reporters = get_new_version_values('reporting-org')

reporters.head()

reporters.tail()

## Other Identifier

others = get_new_version_values('other-identifier.owner-org')

others.head()

others.tail()

## Participating Organizations

participants = get_new_version_values('participating-org')

participants.head()

participants.tail()

## Transaction Providers

transaction_providers = get_new_version_values('transaction.provider-org')

transaction_providers.head()

transaction_providers.tail()

## Transaction Receivers

transaction_receivers = get_new_version_values('transaction.receiver-org')

transaction_receivers.head()

transaction_receivers.tail()

# Check Ref-Names

name_lookup = defaultdict(set)

def get_names_from_organization(item, ref):
    short_names = []

    if '#text' in item:
        short_names.extend([item['#text']])
   
    if 'name' in item:
        get_names_from_organization(item['name'], ref)
   
    if 'narrative' in item:
        narrative = item['narrative']

        if type(narrative) == dict:
            get_names_from_organization(narrative, ref)
        elif type(narrative) == list:
            for subnarrative in narrative:
                get_names_from_organization(subnarrative, ref)
        else:
            short_names.extend([narrative])

    if len(short_names) == 0:
        return

    if ref not in name_lookup:
        name_lookup[ref] = set()

    name_lookup[ref] |= set(short_names)

for organization in organizations.find():
    if 'organisation-identifier' in organization:
        ref = organization['organisation-identifier']

        if type(ref) == dict:
            ref = ref['#text']

        get_names_from_organization(organization, ref)

    if 'reporting-org' in organization:
        reporting_org = organization['reporting-org']

        if '@ref' not in reporting_org:
            continue

        ref = reporting_org['@ref']

        if type(reporting_org) == list:
            for item in reporting_org:
                get_names_from_organization(item, ref)
        elif type(reporting_org) == dict:
            get_names_from_organization(reporting_org, ref)

len(name_lookup)

short_names_by_ref = {}
short_names_by_ref.update(name_lookup)

len(short_names_by_ref)

## Check Ref to Name Mapping

def get_names_from_activity(df):
    for file_id, ref, narrative in zip(df['file_id'], df['ref'], df['narrative']):
        if ref is None or narrative is None:
            continue

        ref = ref.strip()

        if len(ref) == 0:
            continue

        if type(narrative) != list:
            narrative = narrative.strip()

            if len(narrative) > 0:
                name_lookup[ref].add(narrative)

            continue

        for item in narrative:
            if item is None:
                continue

            item = item.strip()

            if len(item) > 0:
                name_lookup[ref].add(item)

get_names_from_activity(reporters)
get_names_from_activity(others)
get_names_from_activity(participants)
get_names_from_activity(transaction_providers)
get_names_from_activity(transaction_receivers)

len(name_lookup)

multiple_names = []

for ref, names in name_lookup.items():
    if len(names) > 1:
        multiple_names.append({'ref': ref, 'narrative': names})

len(multiple_names)

pandas.DataFrame(multiple_names).head()

## Check Name to Ref Mapping

ref_lookup = defaultdict(set)

def get_refs(df):
    for file_id, ref, narrative in zip(df['file_id'], df['ref'], df['narrative']):
        if ref is None or narrative is None:
            continue

        ref = ref.strip()

        if len(ref) == 0:
            continue

        if type(narrative) != list:
            narrative = narrative.strip()

            if len(narrative) > 0:
                ref_lookup[narrative].add(ref)

            continue

        for item in narrative:
            if item is None:
                continue

            item = item.strip()

            if len(item) > 0:
                ref_lookup[item].add(ref)

get_refs(reporters)
get_refs(others)
get_refs(participants)
get_refs(transaction_providers)
get_refs(transaction_receivers)

len(ref_lookup)

multiple_refs = []

for name, refs in ref_lookup.items():
    if len(refs) > 1:
        multiple_refs.append({'narrative': name, 'ref': refs})

len(multiple_refs)

pandas.DataFrame(multiple_refs).head()

## Store as Pickled Files

with open('short_names_by_ref.pickle', 'wb') as f:
    pickle.dump(short_names_by_ref, f)

with open('lookup_by_ref.pickle', 'wb') as f:
    pickle.dump(name_lookup, f)

with open('lookup_by_name.pickle', 'wb') as f:
    pickle.dump(ref_lookup, f)

# Backup Files

!rm -f lookup.tar.gz
!tar -cf lookup.tar iati_urls_by_ref.pickle short_names_by_ref.pickle lookup_by_ref.pickle lookup_by_name.pickle
!gzip lookup.tar

s3_bucket = 'mdang.w210'

!aws s3 cp lookup.tar.gz s3://{s3_bucket}/ --acl public-read
