==> docker-eda/08-check-graph-statistics.ipynb <==
# Install Packages

%matplotlib inline

from __future__ import print_function
from collections import defaultdict
from datetime import datetime
import functools
import io
import itertools
import json
import os
import pandas
import pickle
from pprint import PrettyPrinter
import pymongo
import re
from scipy.sparse import dok_matrix

# Data Check

conn=pymongo.MongoClient('mongodb', 27017)

conn.database_names()

db = conn.iati

activities=db.activities

db.collection_names()

print(activities.count())

# Cache Data

Bring all the activities out of MongoDB and into memory so that we can process them much faster. Iterating over the MongoDB collection takes minutes. Iterating over an in-memory version containing only the fields we're interested in takes seconds.

However, we need a lot of memory to fit everything, so we'll only include a few of the fields which might be useful when constructing the graph.

fields = set([
    '@w210-key',
    'reporting-org',
    'participating-org',
    'recipient-country',
    'transaction'
])

all_activities = []
activities_count = activities.count()

print(datetime.now(), 'Started processing')

for num, activity in enumerate(activities.find()):
    if num % 50000 == 0:
        print(datetime.now(), 'Processed', num, 'of', activities_count)

    activity_copy = { key: value for key, value in activity.items() if key in fields }

    all_activities.append(activity_copy)

print(datetime.now(), 'Finished processing')

# Create Adjacency List Files

def clean_text(text):
    return re.sub('\s+', ' ', text).strip()

def get_text(element, attribute):

    if element is None:
        return None

    if attribute not in element:
        return None

    try:
        value = element[attribute]
    except Exception as e:
        print(element, attribute)
        raise e

    if type(value) == dict:
        if '#text' in value:
            return value['#text']

        return None

    if type(value) != list:
        return value

    return [
        item['#text'] if type(item) == dict and '#text' in item else
            None if type(item) == dict else item
        for item in value
    ]

def get_node_list(parent, field_keys):
    value = parent

    for key in field_keys:

        # If we have a dictionary, we simply access the attribute

        if type(value) == dict:
            if key not in value:
                return []

            value = value[key]
            continue

        # If we have something that is neither a dict nor a list, we
        # cannot navigate further down the JSON object, so we were
        # unable to find what we needed.

        if type(value) != list:
            return []

        # If we have a list, then we'll check the key in each element
        # of the list.

        value = [ item[key] for item in value if item is not None and key in item ]

    if value is None:
        return []

    if type(value) != list:
        value = [value]

    return value

def get_edge_list(
    activity, field_keys, left_element, left_child, right_element, right_child,
    left_reporting_org_fallback, right_reporting_org_fallback):

    node_list = get_node_list(activity, field_keys)

    return_value = []

    left_fallback = None

    if left_reporting_org_fallback is not None:
        left = get_node_list(activity, ['reporting-org'])
        left_fallback = get_text(left[0], left_reporting_org_fallback)

    right_fallback = None

    if right_reporting_org_fallback is not None:
        right = get_node_list(activity, ['reporting-org'])
        right_fallback = get_text(right[0], right_reporting_org_fallback)

    for node in node_list:
        if left_element not in node or right_element not in node:
            continue

        left = get_node_list(node, [left_element])
        right = get_node_list(node, [right_element])

        for left_node, right_node in itertools.product(left, right):
            left_list = get_text(left_node, left_child)
            right_list = get_text(right_node, right_child)

            if left_list is None:
                left_list = left_fallback

            if right_list is None:
                right_list = right_fallback

            if left_list is None or right_list is None:
                continue

            if type(left_list) != list:
                left_list = [left_list]

            if type(right_list) != list:
                right_list = [right_list]

            return_value += [
                (activity['@w210-key'], clean_text(left_value), clean_text(right_value))
                    for left_value, right_value in itertools.product(left_list, right_list)
                        if left_value is not None and right_value is not None
            ]

    return return_value

def get_edges(
    field_path, left_element, left_child, right_element, right_child,
    left_reporting_org_fallback = None, right_reporting_org_fallback = None):

    if field_path is None:
        field_keys = []
    else:
        field_keys = field_path.split('.')

    return_values = []

    for activity in all_activities:
        new_values = get_edge_list(
            activity, field_keys, left_element, left_child, right_element, right_child,
            left_reporting_org_fallback, right_reporting_org_fallback)

        return_values += new_values

    return return_values

## Graph from Root Elements

root_edges_ref = get_edges(
    None, 'reporting-org', '@ref', 'participating-org', '@ref')

len(root_edges_ref)

with open('graph_root_ref.txt', 'wb') as f:
    pickle.dump(root_edges_ref, f)

root_edges_narrative = get_edges(
    None, 'reporting-org', 'narrative', 'participating-org', 'narrative')

len(root_edges_narrative)

with open('graph_root_narrative.txt', 'wb') as f:
    pickle.dump(root_edges_ref, f)

## Graph from Transaction Elements

transaction_edges_ref = get_edges(
    'transaction', 'provider-org', '@ref', 'receiver-org', '@ref', '@ref', None)

len(transaction_edges_ref)

with open('graph_transaction_ref.txt', 'wb') as f:
    pickle.dump(transaction_edges_ref, f)

transaction_edges_narrative = get_edges(
    'transaction', 'provider-org', 'narrative', 'receiver-org', 'narrative', 'narrative', None)

len(transaction_edges_narrative)

with open('graph_transaction_narrative.txt', 'wb') as f:
    pickle.dump(root_edges_ref, f)

## Graph from Recipient Country

recipient_edges_ref = get_edges(
    None, 'recipient-country', '@code', 'reporting-org', '@ref')

len(recipient_edges_ref)

with open('graph_country_ref.txt', 'wb') as f:
    pickle.dump(recipient_edges_ref, f)

# Create Graph Files from Adjacency Lists

We'll want to convert these `.txt` files into actual graphs.

def get_node_id(node_ids, name):

    # Increment the counter if we haven't seen it

    if name not in node_ids:
        node_ids[name] = len(node_ids)

    return node_ids[name]

def load_graph(file_name):
    node_ids = {}

    get_graph_node_id = functools.partial(get_node_id, node_ids)

    # Iterate once in order to initialize the node ID dictionary

    print(datetime.now(), 'Identifying nodes in', file_name)

    with open(file_name, 'rb') as graph_file:
        edges = pickle.load(graph_file)

        for activity_file, source_name, target_name in edges:
            source_id = get_graph_node_id(source_name)
            target_id = get_graph_node_id(target_name)

    node_count = len(node_ids)

    graph_matrix = dok_matrix((node_count, node_count))

    print(datetime.now(), 'Building sparse matrix for', file_name)

    # Iterate again in order to populate the sparse matrix

    with open(file_name, 'rb') as graph_file:
        edges = pickle.load(graph_file)

        for activity_file, source_name, target_name in edges:
            source_id = get_graph_node_id(source_name)
            target_id = get_graph_node_id(target_name)

            graph_matrix[source_id, target_id] += 1

    print(datetime.now(), 'Finished processing', file_name)

    return node_ids, graph_matrix

# Find Dangling Nodes

def get_dangling_nodes(check_axis, graph_ids, graph):

    # Create a reverse lookup table

    graph_names = { value: key for key, value in graph_ids.items() }

    # Sum by the axis and identify the non-zero entries

    sums = graph.sum(axis = check_axis)
    sums = sums.reshape((sums.shape[1 - check_axis], 1))

    return [graph_names[index] for index, value in enumerate(sums) if value == 0]

get_source_nodes = functools.partial(get_dangling_nodes, 0)
get_sink_nodes = functools.partial(get_dangling_nodes, 1)

# Run against All Graphs

for file_name in os.listdir('.'):
    if len(file_name) > 4 and file_name[-4:] == '.ids':
        os.remove(file_name)
    elif len(file_name) > 6 and file_name[-6:] == '.graph':
        os.remove(file_name)

graph_stats = []
graph_data = []

for file_name in os.listdir('.'):
    if file_name[0:5] != 'graph' or file_name[-4:] != '.txt':
        continue

    # Load the file from cache if we've already done the computation once

    if os.path.isfile(file_name + '.ids') and os.path.isfile(file_name + '.graph'):
        print(datetime.now(), 'Loading cached graph for', file_name)

        with open(file_name + '.ids', 'rb') as id_file:
            graph_ids = pickle.load(id_file)

        with open(file_name + '.graph', 'rb') as graph_file:
            graph = pickle.load(graph_file)

    # Otherwise, perform the computation and save the resulting computations

    else:
        print(datetime.now(), 'Loading graph for', file_name)

        graph_ids, graph = load_graph(file_name)

        with open(file_name + '.ids', 'wb') as id_file:
            pickle.dump(graph_ids, id_file)

        with open(file_name + '.graph', 'wb') as graph_file:
            pickle.dump(graph, graph_file)

    source_nodes = get_source_nodes(graph_ids, graph)
    sink_nodes = get_sink_nodes(graph_ids, graph)

    graph_stats.append({
        'graph file': file_name,
        'total nodes': len(graph_ids),
        'has both edges': len(graph_ids) - len(source_nodes) - len(sink_nodes),
        'has only outgoing edges': len(source_nodes),
        'has only incoming edges': len(sink_nodes)
    })

    graph_data.append({
        'file': file_name,
        'graph': graph,
        'all_nodes': graph_ids,
        'source_nodes': source_nodes,
        'sink_nodes': sink_nodes
    })

pandas.DataFrame(graph_stats)
